Refactored BooksList and Book.js components
