+++ src/containers/BooksLists.js
@@ -1,10 +1,7 @@
 import React from 'react';
-import PropTypes from 'prop-types';
-import { connect } from 'react-redux';
 import Book from '../component/Book';
-import { createBook } from '../redux/actions';
 
-const BooksList = ({ books }) => (
+const BooksList = () => (
   <div className="container">
     <table>
       <thead>
@@ -14,32 +11,9 @@ const BooksList = ({ books }) => (
           <th>Category</th>
         </tr>
       </thead>
-      <tbody>
-        {books.map((book) => (
-          <tr key={book.id}>
-            <Book
-              key={book.id}
-              id={book.id}
-              title={book.title}
-              category={book.category}
-            />
-          </tr>
-        ))}
-      </tbody>
+      <tbody><Book /></tbody>
     </table>
   </div>
 );
 
-BooksList.propTypes = {
-  books: PropTypes.arrayOf(PropTypes.object).isRequired,
-};
-
-const mapStateToProps = (state) => ({
-  books: state,
-});
-
-const mapDispatchToProps = (dispatch) => ({
-  createBook: (book) => dispatch(createBook(book)),
-});
-
-export default connect(mapStateToProps, mapDispatchToProps)(BooksList);
+export default BooksList;
